requirements.txt
opencv-python
matplotlib
scipy
# tensorflow_gpu==2.5.3
torchvision
torch
imgaug
numpy
Pillow
nonechucks
scikit_learn
pytorch-lightning
jsonargparse
loguru